Refactor UI folder structure and include web vitals
