+++ application/ui/.gitignore
@@ -41,8 +41,4 @@ build
 # Yarn v2
 .yarn/cache/**
 
-# Stryker
-reports
-.stryker-tmp
-
 src/config.json
