+++ application/ui/package.json
@@ -3,6 +3,7 @@
   "license": "ISC",
   "private": true,
   "scripts": {
+    "analyse": "source-map-explorer 'build/static/js/*.js'",
     "build": "craco build",
     "dependency-check": "yarn depcheck .",
     "dependency-update": "yarn upgrade-interactive",
@@ -30,7 +31,8 @@
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.6.1",
     "redux": "^4.2.0",
-    "typescript": "^4.9.4"
+    "typescript": "^4.9.4",
+    "web-vitals": "^3.1.0"
   },
   "devDependencies": {
     "@craco/craco": "^7.0.0",
@@ -62,6 +64,7 @@
     "prettier-plugin-organize-imports": "^3.2.1",
     "react-scripts": "^5.0.1",
     "sinon": "^15.0.1",
+    "source-map-explorer": "^2.5.3",
     "typedoc": "^0.23.23"
   },
   "browserslist": {
@@ -120,7 +123,8 @@
   },
   "jest": {
     "coveragePathIgnorePatterns": [
-      "src/index.tsx"
+      "src/index.tsx",
+      "src/reportWebVitals.ts"
     ]
   }
 }
