+++ application/ui/src/app/__test__/store.spec.tsx
