+++ application/ui/src/components/homePage/homePage.tsx
@@ -1,6 +1,6 @@
 import { Button } from "nhsuk-react-components";
+import { useAppSelector } from "../../app/hooks";
 import { LOGIN_PATH, MENU_PATH } from "../../constants/paths";
-import { useAppSelector } from "../../hooks";
 import { selectLoggedIn } from "../../slices/authSlice";
 import Layout from "../common/Layout";
