+++ application/ui/src/components/loginPage/loginPage.tsx
@@ -1,10 +1,10 @@
 import { ActionLink } from "nhsuk-react-components";
+import { useAppSelector } from "../../app/hooks";
 import {
 	AUTH_FORGOTTEN_PASSWORD_ACTION_LINK,
 	AUTH_SIGN_UP_ACTION_LINK,
 } from "../../constants/componentIds";
 import { FORGOTTEN_PASSWORD_PATH, REGISTER_PATH } from "../../constants/paths";
-import { useAppSelector } from "../../hooks";
 import { selectError } from "../../slices/loginErrorSlice";
 import { Layout } from "../common";
 import LoginForm from "./loginForm";
