+++ application/ui/src/index.tsx
@@ -3,10 +3,10 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
-import App from "./components/App";
-import { store } from "./store";
+import App from "./App";
+import { store } from "./app/store";
+import reportWebVitals from "./reportWebVitals";
 
-// Stryker disable all
 export const containerElement = document.querySelector(".react-container");
 
 if (!containerElement) {
@@ -28,5 +28,4 @@ root.render(
 	</React.StrictMode>
 );
 
-// TODO: Add in web vitals
-// Stryker restore all
+reportWebVitals();
