+++ application/ui/src/reportWebVitals.ts
@@ -1,0 +1,17 @@
+import { ReportHandler } from "web-vitals";
+
+/* eslint-disable */
+const reportWebVitals = (onPerfEntry?: ReportHandler) => {
+	if (onPerfEntry && onPerfEntry instanceof Function) {
+		import("web-vitals").then(({ getCLS, getFID, getFCP, getLCP, getTTFB }) => {
+			getCLS(onPerfEntry);
+			getFID(onPerfEntry);
+			getFCP(onPerfEntry);
+			getLCP(onPerfEntry);
+			getTTFB(onPerfEntry);
+		});
+	}
+};
+
+export default reportWebVitals;
+/* eslint-enable */
